ekyc-service/app/utils/security.py
from jose import jwt
from fastapi import HTTPException
from app.config import JWT_SECRET

ALGORITHM = "HS256"

def decode_jwt_token(token: str):
    try:
        return jwt.decode(token, JWT_SECRET, algorithms=[ALGORITHM])
    except jwt.ExpiredSignatureError:
        raise HTTPException(status_code=401, detail="Token expired")
    except jwt.InvalidTokenError:
        raise HTTPException(status_code=401, detail="Invalid token")
